requirements.txt
sentence-transformers
langchain
flask
langchain_google_genai
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain_experimental

-e .
